+++ R/Spectra.R
@@ -84,28 +84,21 @@ NULL
 #'   details.
 #'
 #' Data from a `Spectra` object can be **exported** to a file with the `export`
-#' function. The file format for the exported data can be specified with the
-#' parameter `format`; currently only `format = "mgf"` is supported which
-#' exports the data in Mascot Generic Format (MGF). By default all non-empty
-#' spectra variables in `Spectra` are exported.
+#' function. The actual export of the data has to be performed by the `export`
+#' method of the [MsBackend] class defined with parameter `backend` (which
+#' defaults to the backend currently used by the `Spectra` object). Note
+#' however that not all backend classes support export of the data. Please refer
+#' to the help page of the specific backend class to see whether it supports
+#' export of the data and in which file formats.
 #'
 #' The definition of the function is
-#' `export(object, file = tempfile(), format = "mgf",
-#'     variableMapping = spectraVariableMapping(format), ...)` and its
+#' `export(object, backend = object@backend,  ...)` and its
 #' parameters are:
 #'
 #' - `object`: the `Spectra` object to be exported.
 #'
-#' - `file`: the name of the file.
-#'
-#' - `format`: `character(1)` defining the export file format. Currently only
-#'   `format = "mgf"` is supported.
-#'
-#' - `variableMapping`: named `character` defining the mapping of spectra
-#'   variables (names of `variableMapping`) to fields in the output file
-#'   (elements of `variableMapping`). See output of `spectraVariableMapping()`
-#'   for the expected format. All spectra variables for which no mapping to a
-#'   field in the output format is provided are exported as they are.
+#' - `backend`: instance of a class extending [MsBackend] which supports export
+#'   of the data (i.e. which has a defined `export` method).
 #'
 #'
 #' @section Accessing spectra data:
@@ -472,7 +465,8 @@ NULL
 #' @param backend For `Spectra`: [MsBackend-class] to be used as backend. See
 #'     section on creation of `Spectra` objects for details. For `setBackend`:
 #'     instance of [MsBackend-class]. See section on creation of `Spectra`
-#'     objects for details.
+#'     objects for details. For `export`: [MsBackend-class] to be used to export
+#'     the data.
 #'
 #' @param binSize For `bin`: `numeric(1)` defining the size for the m/z bins.
 #'     Defaults to `binSize = 1`.
@@ -508,12 +502,6 @@ NULL
 #'     backends changing this parameter can lead to errors.
 #'     For `combineSpectra`: `factor` defining the grouping of the spectra that
 #'     should be combined. For `lapply`: `factor` how `X` should be splitted.
-#'
-#' @param file For `export`: `character(1)` specifying the of the file to which
-#'     the data should be exported.
-#'
-#' @param format For `export`: `character(1)` defining the format of the output
-#'     file. Currently only `format = "mgf"` is supported.
 #'
 #' @param FUN For `addProcessing`: function to be applied to the peak matrix
 #'     of each spectrum in `object`. For `compareSpectra`: function to compare
@@ -632,12 +620,6 @@ NULL
 #'
 #' @param value replacement value for `<-` methods. See individual
 #'     method description or expected data type.
-#'
-#' @param variableMapping For `export`: a named `character` vector with the
-#'     names of the fields in the exported file to which spectra variables
-#'     should be mapped, names being spectra variable names and elements of the
-#'     vector the fields in the output file. See output from
-#'     `spectraVariableMapping()` for the default.
 #'
 #' @param which for `containsMz`: either `"any"` or `"all"` defining whether any
 #'     (the default) or all provided `mz` have to be present in the spectrum.
@@ -1045,12 +1027,8 @@ setMethod("split", "Spectra", function(x, f, drop = FALSE, ...) {
 #'
 #' @export
 setMethod("export", "Spectra",
-          function(object, file = tempfile(), format = "mgf",
-                   variableMapping = spectraVariableMapping(format), ...) {
-              switch(match.arg(format),
-                     "mgf" = .export_mgf(object, con = file,
-                                         mapping = variableMapping)
-                     )
+          function(object, backend = object@backend, ...) {
+              export(backend, object, ...)
           })
 
 #### ---------------------------------------------------------------------------