+++ NEWS.md
@@ -2,6 +2,9 @@
 
 ## Changes in 0.99.1
 
+- Add an `export,MsBackend` method to allow backends to take care of data
+  export.
+- Refactor `export,Spectra` to use the `MsBackend` class to export the data.
 - Change parameter `source` in `Spectra,character` to `MsBackendMzR` and set
   parameter `backend = source`. Thus by default, the import backend will also
   be used to store the data.