+++ NEWS.md
@@ -1,7 +1,8 @@
 # Spectra 0.99
 
-## Changes in 0.99.1
+## Changes in 0.99.2
 
+- Add `export,MsBackendMzR` to export spectra data to mzML or mzXML file(s).
 - Add an `export,MsBackend` method to allow backends to take care of data
   export.
 - Refactor `export,Spectra` to use the `MsBackend` class to export the data.